refactor(simulation): split StateCycleDetector into focused modules
Extract snapshot logic (creation, serialization, equality) to snapshots.ts
Extract cycle detection algorithm to detection.ts
Extract shared types (BattleUnit, BattleState, getUnitName) to types.ts
Create index.ts barrel export with StateCycleDetector wrapper class
Reduce original StateCycleDetector.ts to facade re-export

All modules under 400 lines:
- types.ts: 31 lines
- snapshots.ts: 137 lines
- detection.ts: 282 lines
- index.ts: 38 lines

Preserves 100% of public API via facade. Zero breaking changes to consumers.
